♻️ refactor: History 구조 분해
